projectiles: Extract all projectiles logic to own module
V2: - spawn_particle(): actually use parameters passed.
    - fix tabs in vcxproj
